+++ src/main/java/roomescape/reservation/ReservationDao.java
@@ -1,4 +1,4 @@
-package roomescape.user.dao;
+package roomescape.reservation;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -8,7 +8,6 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.Reservation;
 import roomescape.time.Time;
 
 @Repository
