+++ src/main/java/roomescape/reservation/dto/ReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.user.dto;
+package roomescape.reservation.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDate;
