+++ src/test/java/roomescape/reservation/ReservationControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.user.controller;
+package roomescape.reservation;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
